dreamerv2/requirements.txt
cloudpickle==3.0.0
crafter==1.8.3
gym==0.12.5
dm_control==1.0.18
jax==0.4.26
jaxlib==0.4.26
numpy==1.24.3
optax==0.2.2
rich==13.7.1
ruamel.yaml==0.18.6
tensorflow[and-cuda]==2.15.1
tensorflow_probability==0.23.0
scikit-image==0.23.2
ffmpeg==1.4
moviepy==1.0.3
imageio==2.34.1
wandb==0.16.6
opencv-python>=4.9,<5.0
shapely==1.8.4
carla==0.9.15
matplotlib==3.8.4
tabulate==0.9.0
